wip: subagent 完了時の自動コミット (5 files)
PDCA kit の SubagentStop hook による自動コミットです。
PR 前に squash して意味のある単位にまとめてください。

A	experiments/01_what_is_rc/run.py
M	src/rc_basics_lab/experiment/__init__.py
M	src/rc_basics_lab/experiment/runner.py
M	src/rc_basics_lab/experiment/split.py
M	src/rc_basics_lab/tasks/mackey_glass.py

PDCA-Kit-Auto-Commit: true

+++ src/rc_basics_lab/tasks/mackey_glass.py
@@ -62,7 +62,7 @@ def _validate(cfg: MackeyGlassConfig) -> None:
 def delay_steps(cfg: MackeyGlassConfig) -> int:
     """``tau / rk4_step``。整数にならない設定は ``ValueError``。"""
     ratio = cfg.tau / cfg.rk4_step
-    steps = int(round(ratio))
+    steps = round(ratio)
     if steps < 1 or abs(ratio - steps) > _RATIO_TOLERANCE * max(1.0, ratio):
         raise ValueError(
             "tau / rk4_step が整数になりません "
